% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chron2prog.R
\name{chron2prog}
\alias{chron2prog}
\title{Transpose load curtailment chronicle to program with begin and end datetimes}
\usage{
chron2prog(tbl_ts, int_step = 1800, char_group = "groupe",
  char_pow = "puissance", char_datetime = "horodate_UTC",
  char_oml = "dmo")
}
\arguments{
\item{tbl_ts}{tibble timeseries which contains load curtailment chronicle}

\item{int_step}{integer value giving step between consecutive datetime values}

\item{char_group}{character string giving group name in timeseries tibble}

\item{char_pow}{character string giving power name in timeseries tibble}

\item{char_datetime}{character string giving datetime name in UTC timezone from timeseries tibble}

\item{char_oml}{character string giving variable name of the offer mobilization leadtime}
}
\value{
a dataframe with three columns : group, begin(included) and end(excluded)
}
\description{
Transpose load curtailment chronicle to program with begin and end datetimes
}
