% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ref.R
\name{create_ref}
\alias{create_ref}
\title{Fonction de calcul des chroniques de référence}
\usage{
create_ref(tbl_cdc, tbl_sites, tbl_eff, tbl_entt, tbl_homol, tbl_indhist,
  tbl_prev, lgl_parallel = FALSE)
}
\arguments{
\item{tbl_cdc}{un tibble dataframe contenant les chroniques réalisées par site et entité}

\item{tbl_sites}{un tibble dataframe contenant les périmètres MA/NEBEF}

\item{tbl_eff}{un tibble dataframe contenant les effacements réalisés pour MA/NEBEF}

\item{tbl_entt}{un tibble dataframe contenant la liste des entités et leurs méthodes respectives}

\item{tbl_homol}{un tibble dataframe contenant la liste des sites homologués aux méthodes HISTORIQUE et PREVISION}

\item{tbl_indhist}{un tibble dataframe contenant les jours d'indisponibilité à la méthode HISTORIQUE par site et entité}

\item{tbl_prev}{un tibble dataframe contenant les chroniques de prévision par site et entité pour la méthode PREVISION}
}
\value{
un nested tibble dataframe contenant les chroniques d'effacement associées à chaque programme d'effacement
}
\description{
Fonction de calcul des chroniques de référence
}
