% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRModeCorrige.R
\name{CRModeCorrige}
\alias{CRModeCorrige}
\title{Title}
\usage{
CRModeCorrige(tbl_cdc, tbl_sites, tbl_eff, tbl_entt, tbl_homol,
  tbl_indhist, tbl_prev)
}
\arguments{
\item{tbl_prev}{}
}
\description{
Title
}
