% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_infos.R
\name{get_infos}
\alias{get_infos}
\title{Title}
\usage{
get_infos(tbl_eff, tbl_entt, tbl_sites, tbl_homol, tbl_indhist)
}
\arguments{
\item{tbl_eff}{}

\item{tbl_entt}{}

\item{tbl_sites}{}

\item{tbl_homol}{}

\item{tbl_indhist}{}
}
\description{
Title
}
