% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_onhold.R
\name{get_onhold}
\alias{get_onhold}
\title{get information about MA/NEBEF data saved in a directory}
\usage{
get_onhold(dir_data)
}
\arguments{
\item{dir_data}{a character string giving the data repository}
}
\value{
a tibble dataframe giving information about load consumption and curtailement timeseries available
}
\description{
get information about MA/NEBEF data saved in a directory
}
