% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadData.R
\name{LoadIndHist}
\alias{LoadIndHist}
\title{Chargement des fichiers des journées d'indisponibilité par entité et site soumis à la méthodes historique aux formats prévues dans les règles SI décrivant les flux en provenance de RTE à destination des GRD}
\usage{
LoadIndHist(dossiers, fichiers = NULL)
}
\arguments{
\item{fichiers}{un vecteur contenant les noms des fichiers des journées d'indisponibilité (facultatif)}

\item{dossier}{le nom du répertoire contenant les fichiers passés en paramètre}
}
\value{
un dataframe comprenant 6 colonnes : MECANISME, CODE_ENTITE, CODE_SITE, DATE
}
\description{
Chargement des fichiers des journées d'indisponibilité par entité et site soumis à la méthodes historique aux formats prévues dans les règles SI décrivant les flux en provenance de RTE à destination des GRD
}
