% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadData.R
\name{LoadListeEntt}
\alias{LoadListeEntt}
\title{Chargement des fichiers de listing des entités aux formats prévues dans les règles SI décrivant les flux en provenance de RTE à destination des GRD}
\usage{
LoadListeEntt(dossiers, fichiers = NULL)
}
\arguments{
\item{fichiers}{un vecteur contenant les noms des fichiers de listing des entités  (facultatif)}

\item{dossier}{le nom du répertoire contenant les fichiers passés en paramètre}
}
\value{
un dataframe comprenant 5 colonnes : MECANISME, CODE_ENTITE, METHODE, DEBUT, FIN
}
\description{
Chargement des fichiers de listing des entités aux formats prévues dans les règles SI décrivant les flux en provenance de RTE à destination des GRD
}
