% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prog2chron.R
\name{prog2chron}
\alias{prog2chron}
\title{Transpose load curtailment program to chronicle with datetime}
\usage{
prog2chron(tbl_prog, int_step = 300, char_group = "groupe",
  char_pow = "puissance", char_begin = "debut", char_end = "fin",
  char_oml = "dmo")
}
\arguments{
\item{tbl_prog}{tibble timeseries which contains load curtailment programs}

\item{int_step}{integer value giving step between consecutive datetime values to return}

\item{char_group}{character string giving group name in programs tibble}

\item{char_pow}{character string giving power name in programs tibble}

\item{char_begin}{character string giving datetime begin name in UTC timezone from programs tibble}

\item{char_end}{character string giving datetime end name in UTC timezone from programs tibble}

\item{char_oml}{character string giving minutes of offer mobilization leadtime from programs tibble}

\item{timezone}{begin and end timezone from programs tibble}
}
\value{
a dataframe with three columns : group, datetime_UTC and pow
}
\description{
Transpose load curtailment program to chronicle with datetime
}
\examples{
library(tidyverse)
library(lubridate)
prog2chron(tbl_prog = ex_tbl.OA, int_step = 300)
}
