% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_eff}
\alias{read_eff}
\title{Chargement des fichiers de programme d'effacements retenus et d'ordres d'activations consolidés en J+3 aux formats prévues dans les règles SI décrivant les flux en provenance de RTE à destination des GRD}
\usage{
read_eff(dossiers, fichiers = NULL)
}
\arguments{
\item{dossiers}{le(s) nom(s) de répertoire contenant les fichiers passés en paramètre}

\item{fichiers}{un vecteur contenant les noms des fichiers de courbes de charges}
}
\value{
un dataframe comprenant 5 colonnes : CODE_ENTITE, DEBUT, FIN, SIGNE, DMO
}
\description{
Chargement des fichiers de programme d'effacements retenus et d'ordres d'activations consolidés en J+3 aux formats prévues dans les règles SI décrivant les flux en provenance de RTE à destination des GRD
}
