% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_homol}
\alias{read_homol}
\title{Chargement des fichiers de listing des sites homologués aux méthodes historique et prévision aux formats prévues dans les règles SI décrivant les flux en provenance de RTE à destination des GRD}
\usage{
read_homol(dossiers, fichiers = NULL)
}
\arguments{
\item{dossiers}{le nom du répertoire contenant les fichiers passés en paramètre}

\item{fichiers}{un vecteur contenant les noms des fichiers de listing des sites homologués (facultatif)}
}
\value{
un dataframe comprenant 6 colonnes : MECANISME, CODE_SITE, METHODE, DEBUT, FIN, VARIANTE
}
\description{
Chargement des fichiers de listing des sites homologués aux méthodes historique et prévision aux formats prévues dans les règles SI décrivant les flux en provenance de RTE à destination des GRD
}
