% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_indispo}
\alias{read_indispo}
\title{Chargement des fichiers des journées d'indisponibilité par entité et site soumis à la méthodes historique aux formats prévues dans les règles SI décrivant les flux en provenance de RTE à destination des GRD}
\usage{
read_indispo(dossiers, fichiers = NULL)
}
\arguments{
\item{dossiers}{le nom du répertoire contenant les fichiers passés en paramètre}

\item{fichiers}{un vecteur contenant les noms des fichiers des journées d'indisponibilité (facultatif)}
}
\value{
un dataframe comprenant 6 colonnes : MECANISME, CODE_ENTITE, CODE_SITE, DATE
}
\description{
Chargement des fichiers des journées d'indisponibilité par entité et site soumis à la méthodes historique aux formats prévues dans les règles SI décrivant les flux en provenance de RTE à destination des GRD
}
