% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_perim}
\alias{read_perim}
\title{Chargement des fichiers de périmètre de flexibilités aux formats prévues dans les règles SI décrivant les flux en provenance de RTE à destination des GRD}
\usage{
read_perim(fichiers = NULL, dossiers)
}
\arguments{
\item{fichiers}{un vecteur contenant les noms des fichiers de périmètre  (facultatif)}

\item{dossier}{le nom du répertoire contenant les fichiers passés en paramètre}
}
\value{
un dataframe comprenant 5 colonnes : DEBUT, FIN, MECANISME, CODE_ENTITE, CODE_SITE, CODE_EIC_GRD, CAPACITE_MAX_H_SITE, CATEGORIE, TYPE_CONTRAT
}
\description{
Chargement des fichiers de périmètre de flexibilités aux formats prévues dans les règles SI décrivant les flux en provenance de RTE à destination des GRD
}
