% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_modcor.R
\name{write_modcor}
\alias{write_modcor}
\title{Fonction d'écriture des chroniques d'effacement ou report au format csv prévu dans les règles MA/NEBEF SI}
\usage{
write_modcor(tbl_chron, char_path)
}
\arguments{
\item{tbl_chron}{un tible dataframe contenant les chroniques d'effacement par site}

\item{char_path}{le répertoire de sauvegarde du fichier généré}
}
\value{
un fichier par mécanisme, semaine et pas de temps au format csv et conforme aux règles SI
}
\description{
Fonction d'écriture des chroniques d'effacement ou report au format csv prévu dans les règles MA/NEBEF SI
}
